+++ CHANGES.rst
@@ -6,6 +6,9 @@ Changelog
 
 New:
 
+- Get email_from_address from registry
+  [kcleong]
+
 - Move the default xml configuration out into xml files. These can be picked up
   by ``i18ndude`` or other po file generators for translations strings. Also
   translate the ``FIELDS_DEFAULT`` fields.